health and SP bars added. values change but things are out of order. text doesnt change
have to fix the text values so they actually update. Also should think about  making the enemy turn into a coroutine so that text displays and slider adjustments happen in order while also blocking the player movement. ughughughugh

THEN we actually have to make the player able to select an enemy to cast a spell on and have that actually do damage/etc.